refactor(database): split sql and postgres adapters
