+++ modules/database/src/adapters/sequelize-adapter/utils/index.ts
@@ -1,4 +1,20 @@
-export * from './populate';
-export * from './query';
+import { MongoIndexOptions, PostgresIndexOptions } from '@conduitplatform/grpc-sdk';
+
 export * from './schema';
 export * from './collectionUtils';
+export function checkIfPostgresOptions(
+  options: MongoIndexOptions | PostgresIndexOptions,
+) {
+  const postgresOptions = [
+    'concurrently',
+    'name',
+    'operator',
+    'parser',
+    'prefix',
+    'unique',
+    'using',
+    'where',
+  ];
+  const result = Object.keys(options).some(option => !postgresOptions.includes(option));
+  return !result;
+}
